requirements.txt
python-telegram-bot
colorama
requests
